Refactoring solution vectors and some extension of number templates (#24)
Solutions are now stored in solution blocks in `SolutionManager` this
should be better for extensibility in the future. Rather than store
mappings in `SolutionManager` they should be accounted for in a separate
class. This also allows for a more plain storage of the solution
vectors.

Also started playing around with gmsh.

closes #21
working on #4